benchmarking: Add client role
The client node will need to send gRPC traffic to Gitaly. We will use
ghz[0] to do this, as it has extensive and well-documented[1] options,
and supports streaming RPCs, unlike `k6s` used by the GPT. To send RPCs
we need the protobuf definition files, so the Gitaly repo must be cloned
on the client as well for reference.

To ensure we are using the same commit on all hosts, we peel the
requested revision after cloning and persist this as a fact that can be
referred to from other hosts. The Go and Ruby versions used by the
designated commit are also parsed from `.tool-versions` and saved for
use on the Gitaly node.

We will also need to copy the `ghz` output over to the Gitaly node for
collection with the other benchmark outputs. In preparation for that,
create a new SSH key and save its pubkey as another fact so we can trust
it on the Gitaly node.

[0] https://github.com/bojand/ghz
[1] https://ghz.sh