benchmarking: Add Ansible config
Ansible can run slowly when performing a large number of operations, and
determining which tasks are slow is difficult with the default output.

Mitigate these issues by enabling pipelining[0], which speeds things up
dramatically and is compatible with the Ubuntu 22.04 hosts we're using,
and the `profile_tasks` callback[1] which print the start time of each
task during execution and a summary of task times on completion.

[0] https://docs.ansible.com/ansible/latest/reference_appendices/config.html#ansible-pipelining
[1] https://docs.ansible.com/ansible/latest/collections/ansible/posix/profile_tasks_callback.html