benchmarking: Add role to create benchmark hosts
Basic benchmarking will require two hosts, a Gitaly instance and a
client instance to send traffic from. Gitaly Cluster is beyond the scope
of this initial effort.

Create a terraform job that creates both hosts, with port 8075 open on
the Gitaly node for traffic. We use a `t2d` instance for Gitaly as these
provide 4 physical cores, as opposed to 2 hyperthreaded cores. In theory
this could reduce performance jitter, though I have not measured this to
be sure.

A disk image containing the test repositories is attached to the Gitaly
node on creation. These repositories are:

- git.git - A smaller repository with a fair amount of history.
- gitlab.git - Uses an object pool and has ~6,000,000 refs.
- linux.git - A large and well-groomed repository.
- homebrew-core.git - Has very large trees.
- chromium.git - Extremely large (40 GiB), with ~4,000,000 refs.

This task borrows its structure from the old Gitaly Cluster demo script in
`_support/terraform`.