benchmarking: Add gitaly role
Build Gitaly itself, along with Gitaly-Ruby and gitlab-shell. The
versions to use for Go, Ruby, and Gitaly are all parsed on the client
node, so that role must be run first.

To support installing arbitrary version of Ruby, we use ruby-build to
compile it from source. To ensure we are using the most relevant Git
version, we build Gitaly with the bundled Git option enabled. Currently
gitlab-shell is built so Gitaly can validate its presence, but is not
used.