benchmarking: Add systemd service for Gitaly
Directly running Gitaly via Ansible as an asynchronous task would be
painful as they require a hard deadline and can be a bit flaky. In
addition, setting resource limits to the correct values would a pain.

To avoid this, create a systemd service for Gitaly so we can use the
`ansible.builtin.systemd` module to control it. Logs can still be
easily retrieved using `journalctl`, and we can directly set resource
limits to match those used in Omnibus GitLab[0].

[0] https://gitlab.com/gitlab-org/omnibus-gitlab/-/blob/73892599/config/templates/runit/runsvdir-start.erb#L20-37