commit: Add option to include tags with `FindCommits` RPC
When requesting FindCommits subsequent requests are needed to get tags
for each commit in the response.

Add an `include_referenced_by` field in FindCommitsRequest and a
corresponding `referenced_by` field in GitCommit that will include refs
for the commits in order to allow clients to avoid having to make
additional requests.