Add OID filtering to ListRefs RPC
Adds a `pointing_at_oids` field to ListRefsRequest which is passed
to `git-for-each-ref` as the `--points-at` option. Since annotated
tags return their own OID as `target`, `peel_tags`/`peeled_target`
fields are added to indicate the target of annotated tags. Note that
`peel_tags` slows down the request by a factor of 3-4.