catfile: Copy over WriteBlob and WriteBlobs helpers
Copy over WriteBlob and WriteBlobs over to the catfile package, as it
cannot use the localrepo helpers MustWriteBlob, MustWriteBlobs due to a
import cycle that would result.

We cannot yet remove the gittest.WriteBlob and gittest.WriteBlobs quite
yet however, as there are still some places that need it in the gittest
package. Namely, TestDeltaIslands, and WriteTree. Once we have
MustWriteTree and MustWriteCommit, then we can also refactor these
functions. At that point, we will be able to delete gittest.WRiteBlob
and gittest.WriteBlobs