repository: Add tests for D/F conflicts in FetchRemote
When fetching refs from a source, there is a potential D/F conflict
which can occur. Consider the following:

1. Source => {refs: ["refs/heads/branch": commit1]}
2. Mirror => {refs: ["refs/heads/branch/conflict": commit2]}

When we pull the refs from the source to the mirror, there is potential
scenario where there is a D/F conflict.

In FetchRemote we prune refs before fetching, unless NoPrune option is
specified. Because of this, there is no D/F conflict occurring by
default.

Add tests to validate this behavior and codify it.