git/stats: Inline `HasBitmap` functionality
When deriving the packfile info for a repository we do this by executing
`HasBitmap()`. This is inefficient as we now have to read the packfiles
directory twice: once via `PackfilesInfoForRepository()` to enumerate
packfiles, and once via `HasBitmap()` to enumerate the bitmap.

Inline `HasBitmap()` into `PackfilesInfoForRepository()` so that we can
fix this inefficiency. All callers of this function were in our test
suite anyway, so we don't really take away an easy way to perform this
check.