git/stats: Compute loose object count more efficiently
In order to calculate the loose object count we're deriving the complete
repository info. This would also contain a bunch of uninteresting info
that makes this a whole lot more expensive than it'd need to be.

Refactor the function to instead use the loose objects info. Note that
it does not matter what cutoff time we pass to the function as the
object count accounts for both stale and non-stale objects anyway.