housekeeping: Implement common helper to test object state
We have three different functions to test for a repository's object
state in different ways. Unify them into a single function that more
thoroughly verifies that the repository is in an expected state. This
helper function can easily be extended going forward to assert more bits
that are relevant in the context of the object database.