housekeeping: Extend test coverage when repacking objects
The tests we have in place for repacking objects are not all that
thorough as they don't really exercise the different configurations that
`RepackObjects()` can be invoked with.

Refactor the test setup to be table-driven and add a bunch of tests to
extend the test coverage. Note that we also explicitly short-circuit the
case where the caller tries to write bitmaps in an incremental repack,
which would cause Git to error out. Having an explicit error for this
case makes it easier to test and ultimately also easier to consume for
an administrator if they were ever to see it in the logs.