Makefile: Deduplicate the version of the Git distribution
Typically, we have up to three different Git versions in Gitaly:

    - Two bundled Git versions that can be toggled with a feature flag.

    - The distributed Git version.

The distributed Git version will always be matching one of the bundled
Git versions, namely the one that is the current default. So let's
deduplicate these versions and just reuse the bundled Git's version so
that we don't need to remember updating the version in multiple places
on minor version bumps.