hooks: Refactor vote hash generation
Currently the `RestoreCustomHooks` RPC generates a vote hash for a
transaction by using the tar archive it receives as input. This means
that if extraction of the archive generates a differing result across
nodes, the generated vote hash will still be the same since it is based
on the preimage of the contents.

This change refactors vote hash generation for custom hooks to instead
walk the extracted `custom_hooks` directory and hash the file name,
permissions, and data. This ensures the vote being cast in the
transaction reflects the actual state of the hooks in the repository.