housekeeping: Construct eager optimization strategy with repo info
Same as with the preceding commit, let's refactor the eager optimization
strategy to get a repository info as input. This allows us to unify the
code paths to derive the repository info.

Note that we're again deleting tests that exercised deriving the info
manually, as this is already covered by tests for derving the repo
info.