housekeeping: Inject repository information via the housekeeping manager
When the caller of `OptimizeRepository()` wants to change the strategy
used to optimize the repository, then they have to derive the repository
information manually. This makes it hard to have a unified code path and
reuse the repository information centrally in the housekeeping manager.

Refactor the code so that callers instead pass in a optimization
strategy constructor. This allows us to inject the centrally-derived
repository information from the housekeeping manager into that
constructor.