Fix TransactionManager still running after tests
TransactionManager's test setup had a race which allowed for the
Run goroutine to be running still after the tests ended. This could
happen as checkManagerError helper was not waiting for either the
test transaction to receive an error or for the Run to return. It
concluded that the manager was still running as soon as the test
transaction was admitted. This is not the case as the manager checks
its context after admitting a transaction. This commit fixes the
unclear situation by waiting for either the manager's Run to really
exit or for the test transaction to be rejected, confirming that the
manager is stilll processing transactions. This should fix the flake
that occurred when the Run goroutine was still running after the tests
concluded.