housekeeping: Construct heuristical optimization strategy with repo info
The heuristical optimization strategy is responsible for deriving the
necessary information it needs all by itself. This makes it hard to
derive this information externally and act on it in other ways. Also, it
makes it harder to see whether constructing the strategy is "pure" or
not given that we also have to pass in the repository info right now.

Convert the constructor to take the repository info as input. Note that
this also allows us to get rid of a lot of duplicate test cases that are
already getting hit when deriving the repository into itself.