hooks: Refactor directory locks
Currently the `RestoreCustomHooks` RPC uses `safe.LockingDirectory` to
prevent concurrent modification of a repository `custom_hooks`
directory. When the `custom_hooks` directory is locked, a `.lock` file
is created inside and subsequent calls to `Lock()` will fail until the
directory is unlocked.

In future commits, having this `.lock` file exist inside the
`custom_hooks` directory will result in vote hashes being constructed
with the lock file. Also it will make atomic directory swap operations
vulnerable to concurrent modification.

To remediate this, indstead of creating a `.lock` file in the
`custom_hooks` directory, a named lock file is created inside the
repository directory. By giving the lock a name the lock is associated
with the `custom_hooks` directory without existing inside of it.