repository: Add test for misdetected Linux license
The linux kernel has a note about user programs that use the kernel
services by normal system calls. This is part of the COPYING file, but
doesn't provide any value on the actual software license.

At the moment this note isn't ignored and FindLicense fails on it. This
change adds a test demonstrating this failure, so we can fix it in the
next commit.

Issue: https://gitlab.com/gitlab-org/gitaly/-/issues/4745