repository: Ignore licenses without a name
The Linux kernel has a note about user programs that use the kernel
services by normal system calls. This is part of the COPYING file, but
doesn't provide any value on the actual software license.
go-license-detector recognizes this note, but it doesn't have further
details on it, because it's not a real license.

This change modifies the code so these notes are ignored.

Changelog: fixed
Issue: https://gitlab.com/gitlab-org/gitaly/-/issues/4745