proto: CommitDiffRequest: Introduce `WhitespaceChanges`
In the CommitDiffRequest message, introduce a new enum
`WhitespaceChanges` and a corresponding field `whitespace_changes`. This
field will allow us to decide between `--ignore-all-space`,
`--ignore-space-change` or no none.

This field also overshadows the earlier field `ignore_whitespace_change`
which was a boolean value and could not be extended. The field
`ignore_whitespace_change` is now marked as deprecated.

Changelog: added