Define new RemoveAll RPC
This RPC is designed to clear a storage before starting a full restore.
The prevents a certain class of bug where if repositories are created
after a backup is taken, then a full restore from backup takes place,
gitlab can no longer creates projects. This is because restore only
removes repositories that it is about to restore, and this eventually
clashes with the repositories created after the backup is taken.