housekeeping: Move `IsPoolRepository()` into `git/stats`
The `housekeeping` package currently hosts the logic to decide whether a
repository is a pool repository or not. This is not a particularly good
fit, but was done in the past to avoid an import cycle. There is another
import cycle coming up though between `git/stats` and `housekeeping`, as
we want to migrate some the information into the `RepositoryInfo`
struct.

Move the functionality into `git/stats` to avoid this import cycle. It's
a much better fit anyway given that this package is responsible for
reporting various different bits about repositories already. So adding
information about whether a repository is an object pool or not feels
natural.