repository: Fix RestoreCustomHooks test depending on the host's umask
In order to calculate the voting hash for transactional updates in the
`RestoreCustomHooks()` RPC we walk all custom hooks and hash their
contents, paths and permission bits. As our tests for this behaviour
expect hardcoded hashes this means that the hashed data must be the same
on every system that runs these tests. But given that we depend on the
permission bits it means that we now implicitly rely on the host's umask
to be the same for every system, or otherwise the tests will fail.

Fix this by setting an explicit umask in the testcase.