housekeeping: Refactor testing metrics for OptimizeRepository
One of our tests for `OptimizeRepository()` exercises whether we're
reporting on specific metrics as expected. This test is quite repetitive
and hard to extend as we expect each testcase to spell out the full
format as expected by `testutil.CollectAndCompare()`.

Refactor the code to instead use an array of `metric` structures from
which we can then compute the expected string.