housekeeping: Enable writing of multi-pack-indices
When writing packfiles, Git also writes an index for every packfile that
gives it the ability to quickly look up objects in that packfile. While
this works great when you have repositories with only a very limited
number of packfiles, these indices start to become insufficient once
repositories have larger numbers of packfiles because Git may have to
scan through every index in order to find a specific object. This is
also a problem for monorepostiories in Gitaly, as we allow for more
packfiles to exist the bigger the repository is.

With the multi-pack index Git has thus introduced a new data structure
to tackle this problem. The MIDX doesn't only span over a single pack,
but instead it typically spans over all packfiles that exist in the
repository. As such, Git does not have to scan through each index
anymore in order to find an object, but only has to consult the MIDX to
find an object.

Furthermore, the MIDX fixes another long-standing issue in Git with
bitmaps. Given an object, bitmaps allow you to find out in constant time
which other objects are transitively reachable from it and are thus an
important optimization especially when serving fetches. But previous to
the introduction of the MIDX, bitmaps could only be written when doing a
full repack of the repository. Doing full repacks is inefficient though,
so consequentially these bitmaps will become stale and thus less able to
speed up specific actions. With the MIDX though it now becomes possible
to also update bitmaps whenever we're doing an incremental repack of
loose objects as the MIDX itself can also have a bitmap.

And last but not least, it is now also possible to have multiple MIDX
bitmaps in a repository and the repositories it links to via the
alternates. This lifts another restriction we had where we couldn't
write bitmaps in repositories connected to an object pool.

So let's enable writing multi-pack-indices when optimizing repositories,
and furthermore let's enable writing bitmaps for pooled repositories
and when doing incremental repacks.

Changelog: performance