housekeeping: Also report inverse metric for data structure existence
When reporting whether a data structure exists or not via a Prometheus
counter we increase the counter for the corresponding "exists" label.
But if no repository ever reports the metric with the inverse value then
clients will never see that line. This makes it harder than necessary to
properly use these metrics e.g. in Grafana.

Improve this by also reporting the inverse label, even though we don't
increase the counter there.