git/stats: Don't stat(3P) all pack directory entries
We're unconditionally statting all directory entries in `objects/pack`
when deriving repository information even though we only need the size
in two specific cases. This means that we waste precious compute cycles
for the majority of files.

Only derive the entry size for those cases where we need it to reduce
the number of syscalls we need to do. While at it, make the case
stricter where a packfile we were about to read has disappeared.
Repository info is only read from `OptimizeRepository()` anyway, and
it's unexpected for any concurrent caller to delete the packfile entry
from that we're about to read.

Changelog: performance