featureflag: Cache for retrieved feature flags
Cache for the feature flags retrieved from the remote.
It runs background job to periodically refresh feature
flags from the remote. The cache is used after the context
is checked. So the feature flags passed with the request
metadata are checked first and cached flags after them and
the final is the default flag value. Now we start with noop
implementation of the cache, no impact. Once we have
implementation for the feature flag provider we can plug
it in and start to RefreshableCache implementation.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/4759