git: Don't have `ConfigPair` implement `Option` interface
When we still used git-config(1) to write Git configuration into our
repositories we relied on `ConfigPair` to implement the `Option`
interface so that we could pass `ConfigPairs` as positional arguments.
We have since stopped writing anything to the Git configuration though
and likely won't start doing so again, so the fact that we can pass a
`ConfigPair` as an `Option` only means that it's easier to get wrong.

Stop implementing the `Option` interface for `ConfigPair`s.