git: Don't pass in Git configuration to `combineArgs()`
We're passing in the `config.GitConfig` from the `ExecCommandFactory` to
`combineArgs()` even though the latter is actually a receiver method on
that type. This is confusing as it makes the reader think that callers
might pass in something different than the factory's Git config.

Refactor the code to instead use the receiver's field.