Remove PostUploadPack RPC
In prior versions, we introduced Sidechannel protocols as an optmization
to avoid transferring huge binary data over gRPC. This protocol was
applied in *Sidechannel version of existing RPCs. We rolled out those
RPCs and replaced all existing clients. There were no traffic to the old
RPCs for a long time. It's time to wipe them out, starting with
PostUploadPack RPC.