housekeeping: Increase number of buckets for size and count metrics
The buckets we have available for the data structure size and count
metrics are quite coarse-grained, which makes it harder to spot
outliers. Furthermore, the cardinality of these labels is not all that
high right now, which means that we definitely have some more room for
adding more buckets.

Increase the bucket sizes to 32 buckets to improve this.