housekeeping: Report recent loose objects
We're reporting both the loose objects count/size and the loose objects
count/size for stale objects. But as the loose objects also include
stale objects the result is that heatmmaps for lookse and loose stale
objects look almost exactly the same as stale objects typically dominate
the recent objects.

Switch the metric to report recent loose objects instead of the overall
loose object count, which is more useful.