remoterepo: Handle SHA256 object hash in `ResolveRevision()`
The `ResolveRevision()` function of the `remoterepo` structure always
parses the object ID returned via the `FindCommit()` RPC call as a SHA1
object hash. This of course doesn't work when handling a repository that
uses SHA256 as object hash.

Fix this by detecting the object hash that is used by the repository.

Changelog: added