objectpool: Fix object format when creating from empty repo
Git is not able to correctly detect the object format used by a source
repository during a clone when the source repository is empty. So in
case the source repository uses SHA256 as object format, the resulting
cloned repository will use the default object hash SHA1 regardless of
that.

This causes a mismatch in object formats for us when we create an object
pool from an empty source repositories, so when we link the pool and its
member together only bad things can happen.

Fix this by explicitly setting the default object format to the one of
the source repository when cloning the repository.

Changelog: fixed