repository: Fetch "remote params" description in FetchRemote tests
Some of the descriptions for `FetchRemote()` tests explicitly mention
that we're testing with "remote params". This is a legacy of the past
where it was possible to fetch either via an on-disk remote that existed
in the Git configuration of that repository or via "remote params" that
were set by the client-side. Nowadays we don't use on-disk remotes at
all anymore though, so clients would always send remote params. As such,
it is redundant to mention that in the description.

Remove mention of "remote params" to make the descriptions less verbose
and more focussed on what matters.