git: Wire up Git v2.39 execution environment
Wire up the Git v2.39 execution environment so that we can toggle
between the v2.38 and v2.39 release tracks. Major new features that
we're interested in include:

    - The upstreamed patch to improve git-receive-pack(1)'s connectivity
      check to not use all local refs as boundary anymore. This is a
      performance improvement that we have already backported into Git
      v2.38.2.gl1.

    - git-merge-tree(1) has learned a new `--stdin` flag to perform a
      series of merges. This may prove useful to to iterate on our
      migration away from libgit2.

    - git-repack(1) has learned to delete packfile bitmaps when writing
      multi-pack-index bitmaps.

    - A deadlock in git-repack(1) when receiving a signal, which we have
      also hit in production systems.

    - A fix where git-patch-id(1) didn't account for changed binary
      files.

There are naturally various other new features, bugfixes and performance
improvements.

The new Git v2.39 binary is protected by a feature flag.

Note that we explicitly test with this feature flag in a single location
only, which is for the `GetPatchID` RPC, as the computed patch ID is
known to change for binaries in Git v2.39 and later.

Changelog: added